assignment1/requirements.txt
groq>=0.9.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
requests>=2.32.0
beautifulsoup4>=4.12.0
streamlit>=1.40.0
python-dotenv>=1.0.0
numpy>=1.26.0
